plotting moved to separate module
