+++ bycycle/burst.py
@@ -2,8 +2,6 @@
 
 import numpy as np
 import pandas as pd
-from scipy.stats import zscore
-import matplotlib.pyplot as plt
 
 from neurodsp.burst import detect_bursts_dual_threshold
 
@@ -169,191 +167,6 @@ def _min_consecutive_cycles(df_shape, n_cycles_min=3):
     df_shape['is_burst'] = is_burst
 
     return df_shape
-
-
-def plot_burst_detect_params(sig, fs, df_shape, osc_kwargs, tlims=None,
-                             figsize=(16, 3), plot_only_result=False):
-    """Create a plot to study how the cycle-by-cycle burst detection
-    algorithm determine bursting periods of a signal.
-
-    Parameters
-    ----------
-    sig : 1d array
-        Time series analyzed to compute `df_shape`.
-    fs : float
-        Sampling rate, in Hz.
-    df_shape : pandas DataFrame
-        Dataframe output of `features.compute_features()`.
-    osc_kwargs : dict
-        Dictionary of thresholds for burst detection used in the function
-        `features.compute_features()` using the kwarg `burst_detection_kwargs`.
-    tlims : tuple of (float, float), optional
-        Start and stop times for plot.
-    figsize : tuple of (float, float), optional
-        Size of figure.
-    plot_only_result : bool, optional, default: False
-        If True, do not plot the subplots showing the parameters.
-
-    Returns
-    -------
-    ax : matplotlib axis handle or list of axis handles
-        If `plot_only_result` = True: return a plot of the burst
-        detection in which periods with bursts are denoted in red.
-
-        If `plot_only_result` = False: return a list of the fig
-        handle followed by the 5 axes.
-
-        In the top plot, the raw signal is plotted in black, and the
-        red line indicates periods defined as oscillatory bursts.
-        The highlighted regions indicate when each burst requirement
-        was violated, color-coded consistently with the plots below.
-
-        - blue: amplitude_fraction_threshold,
-        - red: amplitude_consistency_threshold
-        - yellow: period_consistency_threshold
-        - green: monotonicity_threshold
-    """
-
-    # Normalize signal
-    sig = zscore(sig)
-
-    # Determine time array
-    times = np.arange(0, len(sig) / fs, 1 /fs)
-
-    if tlims is None:
-        tlims = (times[0], times[-1])
-
-    # Determine extrema labels
-    if 'sample_trough' in df_shape.columns:
-        center_e = 'trough'
-        side_e = 'peak'
-    else:
-        center_e = 'peak'
-        side_e = 'trough'
-
-    # Limit to time periods of interest
-    tidx = np.logical_and(times >= tlims[0], times < tlims[1])
-    sig = sig[tidx]
-    times = times[tidx]
-    df_shape = df_shape[(df_shape['sample_last_' + side_e] > int(fs * tlims[0])) &
-                        (df_shape['sample_next_' + side_e] < int(fs * tlims[1]))]
-    df_shape['sample_last_' + side_e] = df_shape['sample_last_' + side_e] - int(fs * tlims[0])
-    df_shape['sample_next_' + side_e] = df_shape['sample_next_' + side_e] - int(fs * tlims[0])
-    df_shape['sample_' + center_e] = df_shape['sample_' + center_e] - int(fs * tlims[0])
-
-    # Determine which samples are defined as bursting
-    is_osc = np.zeros(len(sig), dtype=bool)
-    df_osc = df_shape[df_shape['is_burst']]
-    for _, cyc in df_osc.iterrows():
-        samp_start_burst = cyc['sample_last_' + side_e]
-        samp_end_burst = cyc['sample_next_' + side_e] + 1
-        is_osc[samp_start_burst:samp_end_burst] = True
-
-    if plot_only_result:
-
-        # Plot the time series and indicate peaks and troughs
-        _, ax = plt.subplots(figsize=figsize)
-
-        ax.plot(times, sig, 'k')
-        ax.plot(times[is_osc], sig[is_osc], 'r.')
-        ax.set_xlim(tlims)
-        ax.set_title('Raw z-scored signal. Red trace indicates periods of bursting', size=15)
-        ax.set_xlabel('Time (s)', size=12)
-        ax.set_ylabel('Voltage (normalized)', size=12)
-
-        return ax
-
-    else:
-
-        # Plot the time series and indicate peaks and troughs
-        fig = plt.figure(figsize=(figsize[0], 5*figsize[1]))
-        ax1 = fig.add_subplot(5, 1, 1)
-        ax1.plot(times, sig, 'k')
-        ax1.plot(times[is_osc], sig[is_osc], 'r', linewidth=2)
-        ax1.plot(times[df_shape['sample_' + center_e]], sig[df_shape['sample_' + center_e]],
-                 'm.', ms=10)
-        ax1.plot(times[df_shape['sample_last_' + side_e]], sig[df_shape['sample_last_' + side_e]],
-                 'c.', ms=10)
-        ax1.set_xlim(tlims)
-        ax1.set_xticks([])
-        ax1.set_ylabel('Black: Raw signal\nRed: oscillatory periods')
-        ax1.set_ylim((-4, 4))
-
-        # Highlight where burst detection parameters were violated
-        # Use a different color for each burst detection parameter
-        ax1.fill_between(times[df_shape['sample_last_' + side_e]], -4, 400,
-                         where=df_shape['amp_fraction'] < \
-                            osc_kwargs['amplitude_fraction_threshold'],
-                         interpolate=True, facecolor='blue', alpha=0.5, )
-        ax1.fill_between(times[df_shape['sample_last_' + side_e]], -4, 400,
-                         where=df_shape['amp_consistency'] < \
-                            osc_kwargs['amplitude_consistency_threshold'],
-                         interpolate=True, facecolor='red', alpha=0.5)
-        ax1.fill_between(times[df_shape['sample_last_' + side_e]], -4, 400,
-                         where=df_shape['period_consistency'] < \
-                            osc_kwargs['period_consistency_threshold'],
-                         interpolate=True, facecolor='yellow', alpha=0.5)
-        ax1.fill_between(times[df_shape['sample_last_' + side_e]], -4, 400,
-                         where=df_shape['monotonicity'] < \
-                            osc_kwargs['monotonicity_threshold'],
-                         interpolate=True, facecolor='green', alpha=0.5)
-
-        ax2 = fig.add_subplot(5, 1, 2)
-        ax2.plot(times[df_shape['sample_' + center_e]], df_shape['amp_fraction'], 'k.-')
-        ax2.plot(tlims, [osc_kwargs['amplitude_fraction_threshold'],
-                         osc_kwargs['amplitude_fraction_threshold']], 'k--')
-        ax2.set_xlim(tlims)
-        ax2.set_xticks([])
-        ax2.set_ylim((-.02, 1.02))
-        ax2.set_ylabel('Band amplitude fraction\nthreshold={:.02f}'
-                       .format(osc_kwargs['amplitude_fraction_threshold']))
-        ax2.fill_between(times[df_shape['sample_last_' + side_e]], 0, 100,
-                         where=df_shape['amp_fraction'] < \
-                         osc_kwargs['amplitude_fraction_threshold'],
-                         interpolate=True, facecolor='blue', alpha=0.5)
-
-        ax3 = fig.add_subplot(5, 1, 3)
-        ax3.plot(times[df_shape['sample_' + center_e]], df_shape['amp_consistency'], 'k.-')
-        ax3.plot(tlims, [osc_kwargs['amplitude_consistency_threshold'],
-                         osc_kwargs['amplitude_consistency_threshold']], 'k--')
-        ax3.set_xlim(tlims)
-        ax3.set_xticks([])
-        ax3.set_ylim((-.02, 1.02))
-        ax3.set_ylabel('Amplitude consistency\nthreshold={:.02f}'.
-                       format(osc_kwargs['amplitude_consistency_threshold']))
-        ax3.fill_between(times[df_shape['sample_last_' + side_e]], 0, 100,
-                         where=df_shape['amp_consistency'] < \
-                            osc_kwargs['amplitude_consistency_threshold'],
-                         interpolate=True, facecolor='red', alpha=0.5)
-
-        ax4 = fig.add_subplot(5, 1, 4)
-        ax4.plot(times[df_shape['sample_' + center_e]], df_shape['period_consistency'], 'k.-')
-        ax4.plot(tlims, [osc_kwargs['period_consistency_threshold'],
-                         osc_kwargs['period_consistency_threshold']], 'k--')
-        ax4.set_xlim(tlims)
-        ax4.set_xticks([])
-        ax4.set_ylabel('Period consistency\nthreshold={:.02f}'
-                       .format(osc_kwargs['period_consistency_threshold']))
-        ax4.set_ylim((-.02, 1.02))
-        ax4.fill_between(times[df_shape['sample_last_' + side_e]], 0, 100,
-                         where=df_shape['period_consistency'] < \
-                         osc_kwargs['period_consistency_threshold'],
-                         interpolate=True, facecolor='yellow', alpha=0.5)
-
-        ax5 = fig.add_subplot(5, 1, 5)
-        ax5.plot(times[df_shape['sample_' + center_e]], df_shape['monotonicity'], 'k.-')
-        ax5.plot(tlims, [osc_kwargs['monotonicity_threshold'],
-                         osc_kwargs['monotonicity_threshold']], 'k--')
-        ax5.set_xlim(tlims)
-        ax5.set_ylabel('Monotonicity\nthreshold={:.02f}'
-                       .format(osc_kwargs['monotonicity_threshold']))
-        ax5.set_ylim((-.02, 1.02))
-        ax5.set_xlabel('Time (s)', size=20)
-        ax5.fill_between(times[df_shape['sample_last_' + side_e]], 0, 100,
-                         where=df_shape['monotonicity'] < osc_kwargs['monotonicity_threshold'],
-                         interpolate=True, facecolor='green', alpha=0.5)
-
-        return [fig, ax1, ax2, ax3, ax4, ax5]
 
 
 def detect_bursts_df_amp(df, sig, fs, f_range, amp_threshes=(1, 2),
